+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # RealmJoin Runbooks Changelog 
 
+## 2023-08-14
+* Assign Groups by Template: Improved Error / Edge Case handling
+
 ## 2023-08-11
 * Export Policy Report - Compatibility with Microsoft Graph PowerShell Module 2.x
 * Updated Phone Runbooks